Edit the styling for the final <div> that has a child
Due to the child class appearing alone orinally,
the div.text chain took precedence because this had
more selectors (one of which was a class as well).

By simply specifying the chain div.text.child, we
also allow the child class to have its own precedence
that doesn't affect the previous one.