Introduce a main-container, content, and card-content divs and classes
The main-container div and class houses the 3 main items of header, content, and footer.
The content div and class acts as both a child to main-container and also
as a parent to side-bar and card-content divs and classes.
The card-content div and class is where the different cards will be
housed and styled.